SPOI-8514: Added content in readme. Changed app name. Added lisences.
