--- tutorials/maprapp/src/main/java/com/datatorrent/maprapp/Data.java
@@ -1,41 +1,0 @@
-package com.datatorrent.maprapp;
-
-/**
- * Data: POJO for data
- */
-public class Data
-{
-  private int id;
-  private String name;
-  private String message;
-
-  public int getId()
-  {
-    return id;
-  }
-
-  public void setId(int id)
-  {
-    this.id = id;
-  }
-
-  public String getName()
-  {
-    return name;
-  }
-
-  public void setName(String name)
-  {
-    this.name = name;
-  }
-
-  public String getMessage()
-  {
-    return message;
-  }
-
-  public void setMessage(String message)
-  {
-    this.message = message;
-  }
-}
