+++ tutorials/maprstreams/src/assemble/appPackage.xml
