+++ tutorials/maprstreams/src/main/resources/META-INF/properties.xml
