+++ tutorials/maprstreams/src/test/resources/log4j.properties
